pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "validir"
version = "0.1.0"
authors = [
  { name="Daniel M. Sahu", email="danielmohansahu@gmail.com" },
]
description = "Directory structure template-based validation."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml"
]

[project.scripts]
validir = "validir.cli:main"

[project.urls]
"Homepage" = "https://github.com/danielmohansahu/validir"
"Bug Tracker" = "https://github.com/danielmohansahu/validir/issues"
